Working and tested VPM module; but no SEM-I
This effectively fixes #80, but it will need to be revisited when #67 is implemented (possibly just one function, though, and additional tests).

Note that this might have different behavior than the LKB when * is used as the target. The LKB pops values off a stack (I think), where pyDelphin looks for corresponding values at the same array index on the source side.

This commit also fixes a problem with the unit tests for derivation.py, where warnings sequential DeprecationWarnings were not being caught (because by default only the first instance is raised).