Implement basic SEM-I (read/query only)
This partially implements issue #67. It adds support for loading version 1.0 SEM-I files, querying the contents, and converting it to and from a dict (e.g., for the REST API).

This commit does not include support for hierarchy-aware comparisons of semantic structures, nor does it do any sanity check of the SEM-I itself (e.g. that all variable types, roles, and properties used on predicates are defined in their own sections, etc.)